Refactor Sharing
