+++ talentmap_api/user_profile/models.py
@@ -22,29 +22,6 @@ class UserProfile(models.Model):
         return f"{self.user.username}"
 
 
-class Sharable(models.Model):
-    '''
-    This model represents a shared item from one user to another. The field sharable_model
-    is the string representation of the model, for example 'position.Position', and the
-    sharable_id is the integer representation of the autofield primary key for that model
-    '''
-    sharing_user = models.ForeignKey(UserProfile, related_name="sent_shares")
-    receiving_user = models.ForeignKey(UserProfile, related_name="received_shares")
-
-    # Sharable
-    sharable_id = models.IntegerField(help_text="The ID of the model instance for this sharable")
-    sharable_model = models.TextField(help_text="The string of the model")
-
-    is_read = models.BooleanField(default=False, help_text="Whether this sharable has been read")
-
-    date_created = models.DateTimeField(auto_now_add=True)
-    date_updated = models.DateTimeField(auto_now=True)
-
-    class Meta:
-        managed = True
-        ordering = ["date_created"]
-
-
 class SavedSearch(models.Model):
     '''
     Represents a saved search.
@@ -128,17 +105,3 @@ def post_saved_search_save(sender, instance, created, **kwargs):
     This listener ensures newly created or edited saved searches update their counts
     '''
     instance.update_count()
-
-
-@receiver(post_save, sender=Sharable)
-def post_sharable_save(sender, instance, created, **kwargs):
-    '''
-    This listener notifies the receiving user to notify them of their share.
-    '''
-    if created:
-        # Create a new notification for the receiving user
-        Notification.objects.create(owner=instance.receiving_user,
-                                    message=f"{instance.sharing_user} has shared a {instance.sharable_model.split('.')[-1]} with you")
-
-        # TODO: Add e-mail here when e-mail implementation is determined
-        pass
